M03: retire bitrotted M00 regression trio
The M00-era anchor.py/baseline.py/compare.py CLI was superseded by the
M01/M02 pytest drift gates (test_regression.py::test_anchor_hpv16_drift)
and the M03 multi-seed flow (multi_seed_v3.py + multi_seed_v2.py +
compare_seeds.py + test_m03_short_summary_parity.py). anchor.py was
also broken — its make_sim() passes a dict positionally to hpv.Sim,
which v3 no longer accepts.

  - Extract compute_drift() into tests/regression/drift.py (still used
    by the M01/M02 unit tests).
  - Delete anchor.py, baseline.py, compare.py.
  - Drop the compare.py no-baseline smoke step from the CI workflow.
  - Update tests/regression/README.md: replace the M00-trio narrative
    with a What's-here / pytest-gate table and an M03 multi-seed flow
    section.
