+++ .github/workflows/tests.yaml
@@ -33,9 +33,6 @@ jobs:
       - name: Run all tests
         working-directory: ./tests
         run: pytest test_*.py -n auto --durations=0 --junitxml=test-results.xml # Run actual tests
-      - name: Smoke-check regression compare CLI (no-baseline mode)
-        working-directory: ./tests
-        run: python regression/compare.py
       - name: Publish Test Report
         uses: mikepenz/action-junit-report@v3
         if: always() # always run even if the previous step fails
